API/requirements.txt
Flask
mysql-connector-python
gunicorn
Werkzeug
Flask-Login